Add cloudflare worker compatibility
See/Fixes https://github.com/dnsimple/dnsimple-node/issues/215

In existing code, this lib detects the `fetch` API by checking for `window` - however, there are a few more environments where `fetch` is now the native http handlingmethod and this library could work automatically. Namely cloudflare workers (v8 isolates) and Web/Service Workers.

This PR changes the feature detection to also include WorkerGlobalScope environments, which will detect the above cases.

In addition to the detection change, cloudflare workers appear to hoist `require` calls, so this PR also manually hoists the `require` calls that were previously gated behind an early return in the browser case, but wraps them in a try/catch so that if we incorrectly feature detect node, it will preserve the original error and re-throw it once we've passed the feature gate.